# Conflicts:
#	Panda Player/Controllers/HomeController.cs
#	Panda Player/Controllers/PlaylistsController.cs
#	Panda Player/Controllers/SongsController.cs
#	Panda Player/Models/ViewModels/SongDetailsModel.cs
#	Panda Player/Panda Player.csproj
#	Panda Player/Views/Home/Index.cshtml
#	Panda Player/Views/Shared/Player.cshtml
#	Panda Player/Views/Shared/_Layout.cshtml
#	Panda Player/Views/Shared/_LoginPartial.cshtml
#	Panda Player/Views/User/List.cshtml

+++ Panda Player/Controllers/PlaylistsController.cs
@@ -34,6 +34,11 @@ namespace Panda_Player.Controllers
                 LastPage = lastPage
             };
 
+            if (userPlaylists.Count == 0)
+            {
+                this.AddNotification($"You have not created any playlists yet.", NotificationType.INFO);
+            }
+
             return PartialView(playlistsModel);
         }
 
@@ -131,7 +136,7 @@ namespace Panda_Player.Controllers
             return View(model);
         }
 
-        // GET: Playlists/Edit/5
+        // GET: Playlists/Edit/
         [Authorize]
         public ActionResult Edit(int? id)
         {
@@ -155,7 +160,6 @@ namespace Panda_Player.Controllers
                 return RedirectToAction("MyPlaylists");
             }
 
-
             var model = new PlaylistViewModel
             {
                 Id = playlist.Id,
@@ -166,7 +170,7 @@ namespace Panda_Player.Controllers
             return PartialView(model);
         }
 
-        // POST: Playlists/Edit/5
+        // POST: Playlists/Edit/
         // To protect from overposting attacks, please enable the specific properties you want to bind to, for 
         // more details see https://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
@@ -177,6 +181,12 @@ namespace Panda_Player.Controllers
             if (ModelState.IsValid)
             {
                 var playlistToEdit = db.Playlists.Find(model.Id);
+
+                if (playlistToEdit == null)
+                {
+                    this.AddNotification("Invalid playlist id!", NotificationType.ERROR);
+                    return RedirectToAction("MyPlaylists");
+                }
 
                 playlistToEdit.PlaylistName = model.PlaylistName;
                 playlistToEdit.IsPublic = model.IsPublic;
@@ -191,7 +201,7 @@ namespace Panda_Player.Controllers
             return View(model);
         }
 
-        // GET: Playlists/Delete/5
+        // GET: Playlists/Delete/
         public ActionResult Delete(int? id)
         {
             if (id == null)
@@ -199,7 +209,6 @@ namespace Panda_Player.Controllers
                 this.AddNotification("Invalid playlist id!", NotificationType.ERROR);
                 return RedirectToAction("MyPlaylists");
             }
-
 
             Playlist playlist = db.Playlists.Find(id);
 
@@ -218,16 +227,23 @@ namespace Panda_Player.Controllers
             return View(playlist);
         }
 
-        // POST: Playlists/Delete/5
+        // POST: Playlists/Delete/
         [HttpPost]
         [Authorize]
         public ActionResult DeleteConfirmed(int id)
         {
             Playlist playlist = db.Playlists.Find(id);
+
+            if (playlist == null)
+            {
+                this.AddNotification("Invalid playlist id!", NotificationType.ERROR);
+                return RedirectToAction("MyPlaylists");
+            }
+
             db.Playlists.Remove(playlist);
             db.SaveChanges();
 
-            this.AddNotification("The Playlist has been deleted successfully.", NotificationType.SUCCESS);
+            this.AddNotification($"Playlist '{playlist.PlaylistName}' has been deleted.", NotificationType.INFO);
             return Json(new { Success = true, Url = "/Playlists/MyPlaylists" });
         }
 
@@ -262,12 +278,7 @@ namespace Panda_Player.Controllers
             if (!System.IO.File.Exists($"{myPlayList}"))
             {
                 System.IO.File.Create($"{myPlayList}");
-            }
-
-            if (!System.IO.File.Exists($"{myPlayList}"))
-            {
-                System.IO.File.Create($"{myPlayList}");
-            }
+            }           
 
             System.IO.File.WriteAllText(myPlayList, result.ToString());
         }
@@ -282,6 +293,7 @@ namespace Panda_Player.Controllers
 
             this.AddNotification($"{song.Artist} - {song.Title} has been successfully removed from {playlist.PlaylistName}", NotificationType.WARNING);
             //return Json(new { returnUrl = "Home/Index" });
+
             return RedirectToAction("MyPlaylists");
         }
